browser/components/aboutlogins/tests/chrome/test_login_item.html

<!DOCTYPE HTML>
<html>
<!--
Test the login-item component
-->
<head>
  <meta charset="utf-8">
  <title>Test the login-item component</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <script type="module" src="chrome://browser/content/aboutlogins/components/login-item.js"></script>
  <script type="module" src="chrome://browser/content/aboutlogins/components/confirmation-dialog.js"></script>
  <script src="aboutlogins_common.js"></script>

  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display">
  </p>
<div id="content" style="display: none">
  <iframe id="templateFrame" src="chrome://browser/content/aboutlogins/aboutLogins.html"
          sandbox="allow-same-origin"></iframe>
</div>
<pre id="test">
</pre>
<script>
/** Test the login-item component **/

let gLoginItem, gConfirmationDialog;
const TEST_LOGIN_1 = {
  guid: "123456789",
  origin: "https://example.com",
  username: "user1",
  password: "pass1",
  timeCreated: "1000",
  timePasswordChanged: "2000",
  timeLastUsed: "4000",
};

const TEST_LOGIN_2 = {
  guid: "987654321",
  origin: "https://example.com",
  username: "user2",
  password: "pass2",
  timeCreated: "2000",
  timePasswordChanged: "4000",
  timeLastUsed: "8000",
};

const TEST_BREACH = {
  Name: "Test-Breach",
  breachAlertURL: "https://monitor.firefox.com/breach-details/Test-Breach",
};

const TEST_BREACHES_MAP = new Map();
TEST_BREACHES_MAP.set(TEST_LOGIN_1.guid, TEST_BREACH);

add_task(async function setup() {
  let templateFrame = document.getElementById("templateFrame");
  let displayEl = document.getElementById("display");
  importDependencies(templateFrame, displayEl);

  gLoginItem = document.createElement("login-item");
  displayEl.appendChild(gLoginItem);

  gConfirmationDialog = document.createElement("confirmation-dialog");
  gConfirmationDialog.hidden = true;
  displayEl.appendChild(gConfirmationDialog);
});

add_task(async function test_empty_item() {
  ok(gLoginItem, "loginItem exists");
  is(gLoginItem.shadowRoot.querySelector("input[name='origin']").value, "", "origin should be blank");
  is(gLoginItem.shadowRoot.querySelector("input[name='username']").value, "", "username should be blank");
  is(gLoginItem.shadowRoot.querySelector("input[name='password']").value, "", "password should be blank");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-created")).args.timeCreated, "", "time-created should be blank when undefined");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-changed")).args.timeChanged, "", "time-changed should be blank when undefined");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-used")).args.timeUsed, "", "time-used should be blank when undefined");
});

add_task(async function test_set_login() {
  gLoginItem.setLogin(TEST_LOGIN_1);
  await asyncElementRendered();

  ok(!gLoginItem.dataset.editing, "loginItem should not be in 'edit' mode");
  ok(!gLoginItem.dataset.isNewLogin, "loginItem should not be in 'isNewLogin' mode");
  is(gLoginItem.shadowRoot.querySelector("input[name='origin']").value, TEST_LOGIN_1.origin, "origin should be populated");
  is(gLoginItem.shadowRoot.querySelector("input[name='username']").value, TEST_LOGIN_1.username, "username should be populated");
  is(gLoginItem.shadowRoot.querySelector("input[name='password']").value, TEST_LOGIN_1.password, "password should be populated");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-created")).args.timeCreated, TEST_LOGIN_1.timeCreated, "time-created should be populated");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-changed")).args.timeChanged, TEST_LOGIN_1.timePasswordChanged, "time-changed should be populated");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-used")).args.timeUsed, TEST_LOGIN_1.timeLastUsed, "time-used should be populated");
  let copyButtons = [...gLoginItem.shadowRoot.querySelectorAll(".copy-button")];
  ok(copyButtons.every(button => !isHidden(button)), "The copy buttons should be visible when viewing a login");
});

add_task(async function test_update_breaches() {
  gLoginItem.setLogin(TEST_LOGIN_1);
  gLoginItem.updateBreaches(TEST_BREACHES_MAP);
  await asyncElementRendered();

  let correspondingBreach = TEST_BREACHES_MAP.get(gLoginItem._login.guid);
  let breachAlert = gLoginItem.shadowRoot.querySelector(".breach-alert");
  ok(!breachAlert.hidden, "Breach alert should be visible");
  is(breachAlert.querySelector(".breach-alert-link").href, correspondingBreach.breachAlertURL, "Breach alert link should be equal to the correspondingBreach.breachAlertURL.");
});

add_task(async function test_breach_alert_is_correctly_hidden() {
  gLoginItem.setLogin(TEST_LOGIN_2);
  gLoginItem.updateBreaches(TEST_BREACHES_MAP);
  await asyncElementRendered();

  let breachAlert = gLoginItem.shadowRoot.querySelector(".breach-alert");
  ok(breachAlert.hidden, "Breach alert should not be visible on login without an associated breach.");
});

add_task(async function test_edit_login() {
  gLoginItem.setLogin(TEST_LOGIN_1);
  gLoginItem.shadowRoot.querySelector(".edit-button").click();
  await asyncElementRendered();

  ok(gLoginItem.dataset.editing, "loginItem should be in 'edit' mode");
  ok(isHidden(gLoginItem.shadowRoot.querySelector(".edit-button")), "edit button should be hidden in 'edit' mode");
  ok(!gLoginItem.dataset.isNewLogin, "loginItem should not be in 'isNewLogin' mode");
  let deleteButton = gLoginItem.shadowRoot.querySelector(".delete-button");
  ok(!deleteButton.disabled, "Delete button should be enabled when editing a login");
  is(gLoginItem.shadowRoot.querySelector("input[name='origin']").value, TEST_LOGIN_1.origin, "origin should be populated");
  is(gLoginItem.shadowRoot.querySelector("input[name='username']").value, TEST_LOGIN_1.username, "username should be populated");
  is(gLoginItem.shadowRoot.querySelector("input[name='password']").value, TEST_LOGIN_1.password, "password should be populated");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-created")).args.timeCreated, TEST_LOGIN_1.timeCreated, "time-created should be populated");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-changed")).args.timeChanged, TEST_LOGIN_1.timePasswordChanged, "time-changed should be populated");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-used")).args.timeUsed, TEST_LOGIN_1.timeLastUsed, "time-used should be populated");
  let copyButtons = [...gLoginItem.shadowRoot.querySelectorAll(".copy-button")];
  ok(copyButtons.every(button => isHidden(button)), "The copy buttons should be hidden when editing a login");

  gLoginItem.shadowRoot.querySelector("input[name='username']").value = "newUsername";
  gLoginItem.shadowRoot.querySelector("input[name='password']").value = "newPassword";

  let updateEventDispatched = false;
  document.addEventListener("AboutLoginsUpdateLogin", event => {
    is(event.detail.guid, TEST_LOGIN_1.guid, "event should include guid");
    is(event.detail.origin, TEST_LOGIN_1.origin, "event should include origin");
    is(event.detail.username, "newUsername", "event should include new username");
    is(event.detail.password, "newPassword", "event should include new password");
    updateEventDispatched = true;
  }, {once: true});
  gLoginItem.shadowRoot.querySelector(".save-changes-button").click();
  ok(updateEventDispatched, "Clicking the .save-changes-button should dispatch the AboutLoginsUpdateLogin event");
});

add_task(async function test_edit_login_cancel() {
  gLoginItem.setLogin(TEST_LOGIN_1);
  gLoginItem.shadowRoot.querySelector(".edit-button").click();

  ok(gLoginItem.dataset.editing, "loginItem should be in 'edit' mode");
  is(!!gLoginItem.dataset.isNewLogin, false,
     "loginItem should not be in 'isNewLogin' mode");

  gLoginItem.shadowRoot.querySelector(".cancel-button").click();
  gConfirmationDialog.shadowRoot.querySelector(".confirm-button").click();

  await SimpleTest.promiseWaitForCondition(
    () => gConfirmationDialog.hidden,
    "waiting for confirmation dialog to hide"
  );

  ok(!gLoginItem.dataset.editing, "loginItem should not be in 'edit' mode");
  ok(!gLoginItem.dataset.isNewLogin, "loginItem should not be in 'isNewLogin' mode");
});

add_task(async function test_reveal_password_change_selected_login() {
  gLoginItem.setLogin(TEST_LOGIN_1);
  let revealCheckbox = gLoginItem.shadowRoot.querySelector(".reveal-password-checkbox");
  let passwordInput = gLoginItem.shadowRoot.querySelector("input[name='password']");

  ok(!revealCheckbox.checked, "reveal-checkbox should not be checked by default");
  is(passwordInput.type, "password", "Password should be masked by default");
  revealCheckbox.click();
  ok(revealCheckbox.checked, "reveal-checkbox should be checked after clicking");
  await SimpleTest.promiseWaitForCondition(() => passwordInput.type == "text",
    "waiting for password input type to change after checking for master password");
  is(passwordInput.type, "text", "Password should be unmasked when checkbox is clicked");

  gLoginItem.setLogin(TEST_LOGIN_2);
  ok(!revealCheckbox.checked, "reveal-checkbox should be unchecked when changing logins");
  is(passwordInput.type, "password", "Password should be masked by default when switching logins");
});

add_task(async function test_set_login_empty() {
  gLoginItem.setLogin({});
  await asyncElementRendered();

  ok(gLoginItem.dataset.editing, "loginItem should be in 'edit' mode");
  ok(isHidden(gLoginItem.shadowRoot.querySelector(".edit-button")), "edit button should be hidden in 'edit' mode");
  ok(gLoginItem.dataset.isNewLogin, "loginItem should be in 'isNewLogin' mode");
  let deleteButton = gLoginItem.shadowRoot.querySelector(".delete-button");
  ok(deleteButton.disabled, "Delete button should be disabled when creating a login");
  is(gLoginItem.shadowRoot.querySelector("input[name='origin']").value, "", "origin should be empty");
  is(gLoginItem.shadowRoot.querySelector("input[name='username']").value, "", "username should be empty");
  is(gLoginItem.shadowRoot.querySelector("input[name='password']").value, "", "password should be empty");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-created")).args.timeCreated, "", "time-created should be blank when undefined");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-changed")).args.timeChanged, "", "time-changed should be blank when undefined");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-used")).args.timeUsed, "", "time-used should be blank when undefined");
  let copyButtons = [...gLoginItem.shadowRoot.querySelectorAll(".copy-button")];
  ok(copyButtons.every(button => isHidden(button)), "The copy buttons should be hidden when creating a login");

  let createEventDispatched = false;
  document.addEventListener("AboutLoginsCreateLogin", event => {
    createEventDispatched = true;
  }, {once: true});
  gLoginItem.shadowRoot.querySelector(".save-changes-button").click();
  ok(!createEventDispatched, "Clicking the .save-changes-button shouldn't dispatch the event when fields are invalid");
  let originInput = gLoginItem.shadowRoot.querySelector("input[name='origin']");
  ok(originInput.matches(":invalid"), "origin value is required");
  is(originInput.value, "", "origin input should be blank at start");

  originInput.focus();
  for (let originTuple of [
    ["ftp://ftp.example.com/", "ftp://ftp.example.com/"],
    ["https://example.com/", "https://example.com/"],
    ["http://example.com/", "http://example.com/"],
    ["www.example.com/bar", "https://www.example.com/bar"],
    ["example.com/foo", "https://example.com/foo"],
  ]) {
    originInput.value = originTuple[0];
    sendKey("TAB");
    is(originInput.value, originTuple[1],
      "origin input should have https:// prefix when not provided by user");
    // Return focus back to the origin input
    synthesizeKey("VK_TAB", { shiftKey: true });
  }

  gLoginItem.shadowRoot.querySelector("input[name='username']").value = "user1";
  gLoginItem.shadowRoot.querySelector("input[name='password']").value = "pass1";

  document.addEventListener("AboutLoginsCreateLogin", event => {
    is(event.detail.guid, undefined, "event should not include guid");
    is(event.detail.origin, "https://example.com/foo", "event should include origin");
    is(event.detail.username, "user1", "event should include new username");
    is(event.detail.password, "pass1", "event should include new password");
    createEventDispatched = true;
  }, {once: true});
  gLoginItem.shadowRoot.querySelector(".save-changes-button").click();
  ok(createEventDispatched, "Clicking the .save-changes-button should dispatch the AboutLoginsCreateLogin event");
});

add_task(async function test_different_login_modified() {
  gLoginItem.setLogin(TEST_LOGIN_1);
  let otherLogin = Object.assign({}, TEST_LOGIN_1, {username: "fakeuser", guid: "fakeguid"});
  gLoginItem.loginModified(otherLogin);
  await asyncElementRendered();

  is(gLoginItem.shadowRoot.querySelector("input[name='origin']").value, TEST_LOGIN_1.origin, "origin should be unchanged");
  is(gLoginItem.shadowRoot.querySelector("input[name='username']").value, TEST_LOGIN_1.username, "username should be unchanged");
  is(gLoginItem.shadowRoot.querySelector("input[name='password']").value, TEST_LOGIN_1.password, "password should be unchanged");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-created")).args.timeCreated, TEST_LOGIN_1.timeCreated, "time-created should be unchanged");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-changed")).args.timeChanged, TEST_LOGIN_1.timePasswordChanged, "time-changed should be unchanged");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-used")).args.timeUsed, TEST_LOGIN_1.timeLastUsed, "time-used should be unchanged");
});

add_task(async function test_different_login_removed() {
  gLoginItem.setLogin(TEST_LOGIN_1);
  let otherLogin = Object.assign({}, TEST_LOGIN_1, {username: "fakeuser", guid: "fakeguid"});
  gLoginItem.loginRemoved(otherLogin);
  await asyncElementRendered();

  is(gLoginItem.shadowRoot.querySelector("input[name='origin']").value, TEST_LOGIN_1.origin, "origin should be unchanged");
  is(gLoginItem.shadowRoot.querySelector("input[name='username']").value, TEST_LOGIN_1.username, "username should be unchanged");
  is(gLoginItem.shadowRoot.querySelector("input[name='password']").value, TEST_LOGIN_1.password, "password should be unchanged");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-created")).args.timeCreated, TEST_LOGIN_1.timeCreated, "time-created should be unchanged");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-changed")).args.timeChanged, TEST_LOGIN_1.timePasswordChanged, "time-changed should be unchanged");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-used")).args.timeUsed, TEST_LOGIN_1.timeLastUsed, "time-used should be unchanged");
});

add_task(async function test_login_modified() {
  gLoginItem.setLogin(TEST_LOGIN_1);
  let modifiedLogin = Object.assign({}, TEST_LOGIN_1, {username: "updateduser"});
  gLoginItem.loginModified(modifiedLogin);
  await asyncElementRendered();

  is(gLoginItem.shadowRoot.querySelector("input[name='origin']").value, modifiedLogin.origin, "origin should be updated");
  is(gLoginItem.shadowRoot.querySelector("input[name='username']").value, modifiedLogin.username, "username should be updated");
  is(gLoginItem.shadowRoot.querySelector("input[name='password']").value, modifiedLogin.password, "password should be updated");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-created")).args.timeCreated, modifiedLogin.timeCreated, "time-created should be updated");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-changed")).args.timeChanged, modifiedLogin.timePasswordChanged, "time-changed should be updated");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-used")).args.timeUsed, modifiedLogin.timeLastUsed, "time-used should be updated");
});

add_task(async function test_login_removed() {
  gLoginItem.setLogin(TEST_LOGIN_1);
  gLoginItem.loginRemoved(TEST_LOGIN_1);
  await asyncElementRendered();

  is(gLoginItem.shadowRoot.querySelector("input[name='origin']").value, "", "origin should be cleared");
  is(gLoginItem.shadowRoot.querySelector("input[name='username']").value, "", "username should be cleared");
  is(gLoginItem.shadowRoot.querySelector("input[name='password']").value, "", "password should be cleared");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-created")).args.timeCreated, "", "time-created should be blank when undefined");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-changed")).args.timeChanged, "", "time-changed should be blank when undefined");
  is(document.l10n.getAttributes(gLoginItem.shadowRoot.querySelector(".time-used")).args.timeUsed, "", "time-used should be blank when undefined");
});

</script>

</body>
</html>
